[KMS] move util from envelope to kms package
